Delete tag store
