--- lib/assets/javascripts/new-dashboard/store/actions/shared.js
@@ -1,3 +1,0 @@
-export function setResultsPerPage (context, perPage) {
-  context.commit('setResultsPerPage', perPage);
-}
