+++ lib/assets/javascripts/new-dashboard/store/actions/visualizations.js
@@ -99,6 +99,10 @@ export function resetFilters (DEFAULT_VALUES) {
   };
 }
 
+export function setResultsPerPage (context, perPage) {
+  context.commit('setResultsPerPage', perPage);
+}
+
 export function setURLOptions (context, options) {
   context.commit('setPagination', parseInt(options.page || 1));
   context.commit('setFilterType', options.filter);
