+++ lib/assets/javascripts/new-dashboard/store/modules/datasets/index.js
@@ -1,7 +1,5 @@
 import Filters from 'new-dashboard/core/configuration/filters';
 
-import * as SharedActions from '../../actions/shared';
-import * as SharedMutations from '../../mutations/shared';
 import * as VisualizationActions from '../../actions/visualizations';
 import * as VisualizationMutations from '../../mutations/visualizations';
 
@@ -28,12 +26,12 @@ const datasets = {
     resultsPerPage: 1
   },
   mutations: {
-    setFetchingState: SharedMutations.setFetchingState,
+    setFetchingState: VisualizationMutations.setFetchingState,
     setFilterType: VisualizationMutations.setFilterType,
     setOrder: VisualizationMutations.setOrder(DEFAULT_VALUES),
-    setPagination: SharedMutations.setPagination,
-    setResultsPerPage: SharedMutations.setResultsPerPage,
-    setRequestError: SharedMutations.setRequestError,
+    setPagination: VisualizationMutations.setPagination,
+    setResultsPerPage: VisualizationMutations.setResultsPerPage,
+    setRequestError: VisualizationMutations.setRequestError,
     setVisualizations: VisualizationMutations.setVisualizations,
     updateNumberLikes: VisualizationMutations.updateNumberLikes,
     updateVisualization: VisualizationMutations.updateVisualization
@@ -44,7 +42,7 @@ const datasets = {
     like: VisualizationActions.like,
     order: VisualizationActions.order,
     resetFilters: VisualizationActions.resetFilters(DEFAULT_VALUES),
-    setResultsPerPage: SharedActions.setResultsPerPage,
+    setResultsPerPage: VisualizationActions.setResultsPerPage,
     setURLOptions: VisualizationActions.setURLOptions,
     updateVisualization: VisualizationActions.updateVisualization,
