+++ lib/assets/javascripts/new-dashboard/store/modules/maps/index.js
@@ -2,8 +2,6 @@
 import featuredFavoritedMaps from 'new-dashboard/store/modules/maps/featured-favorited-maps';
 import Filters from 'new-dashboard/core/configuration/filters';
 
-import * as SharedActions from '../../actions/shared';
-import * as SharedMutations from '../../mutations/shared';
 import * as VisualizationActions from '../../actions/visualizations';
 import * as VisualizationMutations from '../../mutations/visualizations';
 
@@ -34,12 +32,12 @@ const maps = {
   },
   getters: {},
   mutations: {
-    setFetchingState: SharedMutations.setFetchingState,
+    setFetchingState: VisualizationMutations.setFetchingState,
     setFilterType: VisualizationMutations.setFilterType,
     setOrder: VisualizationMutations.setOrder(DEFAULT_VALUES),
-    setPagination: SharedMutations.setPagination,
-    setResultsPerPage: SharedMutations.setResultsPerPage,
-    setRequestError: SharedMutations.setRequestError,
+    setPagination: VisualizationMutations.setPagination,
+    setResultsPerPage: VisualizationMutations.setResultsPerPage,
+    setRequestError: VisualizationMutations.setRequestError,
     setVisualizations: VisualizationMutations.setVisualizations,
     updateNumberLikes: VisualizationMutations.updateNumberLikes,
     updateVisualization: VisualizationMutations.updateVisualization
@@ -50,7 +48,7 @@ const maps = {
     like: VisualizationActions.like,
     order: VisualizationActions.order,
     resetFilters: VisualizationActions.resetFilters(DEFAULT_VALUES),
-    setResultsPerPage: SharedActions.setResultsPerPage,
+    setResultsPerPage: VisualizationActions.setResultsPerPage,
     setURLOptions: VisualizationActions.setURLOptions,
     updateVisualization: VisualizationActions.updateVisualization,
