+++ lib/assets/javascripts/new-dashboard/store/modules/recent-content/index.js
@@ -1,7 +1,6 @@
 import { defaultParams as filtersDefaultParams } from 'new-dashboard/core/configuration/filters';
 import toObject from 'new-dashboard/utils/to-object';
 
-import * as SharedMutations from '../../mutations/shared';
 import * as VisualizationActions from '../../actions/visualizations';
 import * as VisualizationMutations from '../../mutations/visualizations';
 
@@ -22,8 +21,8 @@ const recentContent = {
     }
   },
   mutations: {
-    setFetchingState: SharedMutations.setFetchingState,
-    setRequestError: SharedMutations.setRequestError,
+    setFetchingState: VisualizationMutations.setFetchingState,
+    setRequestError: VisualizationMutations.setRequestError,
     updateNumberLikes: VisualizationMutations.updateNumberLikes,
     updateVisualization: VisualizationMutations.updateVisualization,
