+++ lib/assets/javascripts/new-dashboard/store/modules/search/index.js
@@ -1,6 +1,5 @@
 import toObject from 'new-dashboard/utils/to-object';
 import * as VisualizationActions from '../../actions/visualizations';
-import * as TagsMutations from '../../mutations/tags';
 
 const search = {
   namespaced: true,
@@ -84,7 +83,11 @@ const search = {
       state.datasets.isFetching = false;
     },
     setTags (state, tags) {
-      TagsMutations.setTags(state, tags);
+      state.tags.results = tags.result;
+      state.tags.numResults = tags.total;
+      state.tags.numPages = Math.ceil(tags.total / state.resultsPerPage);
+
+      state.tags.isFetching = false;
     },
     updateNumberLikes () {
       // NOOP. This method avoids action not finding mutation.
