--- lib/assets/javascripts/new-dashboard/store/mutations/tags.js
@@ -1,8 +1,0 @@
-
-export function setTags (state, tagsData) {
-  state.tags.results = tagsData.result;
-  state.tags.numResults = tagsData.total;
-  state.tags.numPages = Math.ceil(tagsData.total / state.resultsPerPage);
-
-  state.tags.isFetching = false;
-}
