+++ lib/assets/javascripts/new-dashboard/store/mutations/visualizations.js
@@ -1,7 +1,18 @@
 import toObject from 'new-dashboard/utils/to-object';
 
+export function setFetchingState (state) {
+  state.isFetching = true;
+  state.isErrored = false;
+  state.error = {};
+}
+
 export function setFilterType (state, filterType = 'mine') {
   state.filterType = filterType;
+}
+
+export function setPagination (state, page = 1) {
+  state.page = page;
+  state.numPages = Math.ceil(state.metadata.total_entries / state.resultsPerPage) || 1;
 }
 
 export function setOrder (DEFAULT_VALUES) {
@@ -9,6 +20,16 @@ export function setOrder (DEFAULT_VALUES) {
     state.order = orderOptions.order || DEFAULT_VALUES.order;
     state.orderDirection = orderOptions.direction || DEFAULT_VALUES.orderDirection;
   };
+}
+
+export function setResultsPerPage (state, resultsPerPage = 12) {
+  state.resultsPerPage = resultsPerPage;
+}
+
+export function setRequestError (state, error) {
+  state.isFetching = false;
+  state.isErrored = true;
+  state.error = error;
 }
 
 export function setVisualizations (state, visualizationsData) {
